LLM_ChatBot/requirements.txt
transformers
torch
huggingface_hub
flask
flask_cors
python-dotenv
